Fix some cast AI starting dead after completing a level
The persistant data was always written / read for the first 8 clients
(MAX_COOP_CLIENTS). However if sv_maxCoopClients is less than 8, cast
AI clients are in the remaining slots. This means that health, weapon,
etc are saved for some cast AI clients when a level is completed. If
cast AI had 0 health, the cast AI will be dead in the next level.